requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
scipy>=1.11.0
einops>=0.7.0
omegaconf>=2.3.0
pytorch-lightning>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
diffusers>=0.25.0
safetensors>=0.3.0
timm>=0.9.0
kornia>=0.7.0
segment-anything>=1.0.0
git+https://github.com/IDEA-Research/DWPose.git
git+https://github.com/facebookresearch/dinov2.git